refactor(sports): extract sortContracts to shared web/lib utility
